change to 64bit block id. see http://rb.corp.taobao.com/r/57532/
git-svn-id: http://code.taobao.org/svn/tfs@2104 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-tmp-trunk/src/message/block_info_message_v2.h
@@ -347,11 +347,11 @@ namespace tfs
         virtual int serialize(common::Stream& output)  const ;
         virtual int deserialize(common::Stream& input);
         virtual int64_t length() const;
-        inline void set_block_id(const uint32_t block_id)
+        inline void set_block_id(const uint64_t block_id)
         {
           block_id_ = block_id;
         }
-        inline uint32_t get_block_id() const
+        inline uint64_t get_block_id() const
         {
           return block_id_;
         }
